refactor state a bit
