+++ app/js/app.js
@@ -9,20 +9,13 @@ import { createShip } from '-/player/ship'
 import { createBasicUI } from '-/ui/ui'
 import { createControls } from '-/controls/controls'
 import logger from './logger'
-import { addMessage } from '-/state'
-import state from '-/state'
+import sceneState, { addMessage } from '-/state/branches/scene'
+import { toggleConsole, toggleHelp } from '-/state/branches/gui'
 
 /**
  * Global App State
  */
 const Void = {
-  players: [],
-  player: {
-    ship: null
-  },
-  time: { value: 100000000 },
-  scene: null,
-  world: null,
   animateCallbacks: []
 }
 
@@ -52,7 +45,7 @@ const main = async () => {
 
   logger.debug(addMessage('init: creating player ship...'))
   const { ship, animate: animateShip } = await createShip()
-  state.set([ 'scene', 'player', 'id' ], ship.uuid)
+  sceneState.set([ 'player', 'id' ], ship.uuid)
 
   logger.debug(addMessage('init: adding ship and camera to scene...'))
   ship.add(camera)
@@ -63,14 +56,8 @@ const main = async () => {
   const controls = createControls(
     { type: 'fly', ship, camera, scene },
     {
-      toggleConsole: debounce(100, () => {
-        const consoleState = state.select([ 'gui', 'console' ])
-        consoleState.set('hidden', !consoleState.get('hidden'))
-      }),
-      toggleHelp: debounce(100, () => {
-        const helpState = state.select([ 'gui', 'help' ])
-        helpState.set('hidden', !helpState.get('hidden'))
-      })
+      toggleConsole: debounce(100, toggleConsole),
+      toggleHelp: debounce(100, toggleHelp)
     }
   )
 
@@ -78,7 +65,7 @@ const main = async () => {
   addAnimateCallback(animateShip)
   addAnimateCallback(delta => controls.update(delta))
   addAnimateCallback(() => {
-    state.set([ 'scene', 'bodyCount' ], physics.system.bodies.length)
+    sceneState.set('bodyCount', physics.system.bodies.length)
   })
 
   logger.debug(addMessage('init: creating dat.gui elements...'))
