+++ app/js/controls/fly-controls.js
@@ -1,7 +1,7 @@
 import Keyboard from 'syncinput/Keyboard'
 import Mouse from 'syncinput/Mouse'
-import state from '-/state'
-import { setSelected } from '-/state'
+import sceneState, { setSelected } from '-/state/branches/scene'
+import guiState from '-/state/branches/gui'
 import { calcObjectDistance, calcDistances } from './utils'
 
 const defaultMap = {
@@ -55,7 +55,7 @@ export default class KeyboardControls {
     )
     const newSpeed = this.movementSpeed + dir * step
     this.movementSpeed = newSpeed
-    state.set([ 'scene', 'player', 'movementSpeed' ], newSpeed)
+    sceneState.set([ 'player', 'movementSpeed' ], newSpeed)
   }
 
   update (delta) {
@@ -82,7 +82,7 @@ export default class KeyboardControls {
       setSelected(null)
     }
 
-    if (state.get([ 'gui', 'console', 'hidden' ]) === true) {
+    if (guiState.get([ 'console', 'hidden' ]) === true) {
       Object.keys(this.keymap.moveState).map(key => {
         if (this.keyboard.keyPressed(key)) {
           this.moveState[this.keymap.moveState[key]] = 1
