+++ app/js/state/branches/scene.js
@@ -1,0 +1,14 @@
+import state from '-/state'
+
+const sceneState = state.select('scene')
+
+export const addMessage = msg => {
+  sceneState.apply('messages', messages => [].concat(messages, msg))
+  return msg
+}
+
+export const setSelected = data => {
+  sceneState.set('selected', data ? JSON.stringify(data, null, 2) : '')
+}
+
+export default sceneState
