+++ app/js/state/index.js
@@ -1,0 +1,3 @@
+import state from './state'
+
+export default state
